backend/requirements.txt
fastapi
uvicorn
pydantic
mkdocs
mkdocs-material
pytest
python-jose
httpx
passlib
clerk-backend-api===3.0.3
python-dotenv

